Compilation fixed

+++ Models/Event/Event.cpp
@@ -1,6 +1,6 @@
 #include "Event.hpp"
 
-Event::Event(int id, std::string location){
+Event::Event(long id, std::string location){
 	this->id = id;
 	this->eventLocation = location;
 }
